Cleaned up stores from transition implementations
