+++ fedn/network/storage/statestore/stores/shared.py
@@ -1,8 +1,2 @@
-def from_document(document: dict) -> dict:
-    document["id"] = str(document["_id"])
-    del document["_id"]
-    return document
-
-
 class EntityNotFound(Exception):
     pass
