+++ fedn/network/storage/statestore/stores/store.py
@@ -1,19 +1,28 @@
+import uuid
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Generic, List, Tuple, TypeVar
+from datetime import datetime
+from typing import Any, Dict, Generic, List, Tuple, Type, TypeVar, Union
 
 import pymongo
-from bson import ObjectId
 from pymongo.database import Database
-
-from fedn.network.storage.statestore.stores.shared import from_document
+from sqlalchemy import func, select
+from sqlalchemy.orm import Session as SessionClass
 
 T = TypeVar("T")
 
 
+def from_document(document: dict) -> dict:
+    del document["_id"]
+    return document
+
+
 class Store(ABC, Generic[T]):
+    """Abstract class for a store."""
+
     @abstractmethod
     def get(self, id: str) -> T:
-        """Get an entity by id
+        """Get an entity by id.
+
         param id: The id of the entity
             type: str
         return: The entity
@@ -21,17 +30,20 @@ class Store(ABC, Generic[T]):
         pass
 
     @abstractmethod
-    def update(self, id: str, item: T) -> Tuple[bool, Any]:
-        """Update an existing entity
+    def update(self, item: T) -> Tuple[bool, Union[T, str]]:
+        """Update an existing entity.
+
+        Will do a patch if fields in T are left unset
         param item: The entity to update
-            type: Client
+            type: T
         return: A tuple with a boolean and a message if failure or the updated entity if success
         """
         pass
 
     @abstractmethod
-    def add(self, item: T) -> Tuple[bool, Any]:
-        """Add an entity
+    def add(self, item: T) -> Tuple[bool, Union[T, str]]:
+        """Add an entity.
+
         param item: The entity to add
               type: T
         return: A tuple with a boolean and a message if failure or the added entity if success
@@ -40,7 +52,8 @@ class Store(ABC, Generic[T]):
 
     @abstractmethod
     def delete(self, id: str) -> bool:
-        """Delete an entity
+        """Delete an entity.
+
         param id: The id of the entity
             type: str
         return: A boolean indicating success or failure
@@ -48,69 +61,9 @@ class Store(ABC, Generic[T]):
         pass
 
     @abstractmethod
-    def list(
-        self,
-        limit: int,
-        skip: int,
-        sort_key: str,
-        sort_order=pymongo.DESCENDING,
-        **kwargs,
-    ) -> Dict[int, List[T]]:
-        pass
+    def list(self, limit: int = 0, skip: int = 0, sort_key: str = None, sort_order=pymongo.DESCENDING, **kwargs) -> List[T]:
+        """Select entities.
 
-    @abstractmethod
-    def count(self, **kwargs) -> int:
-        pass
-
-
-class MongoDBStore(Store[T], Generic[T]):
-    def __init__(self, database: Database, collection: str):
-        self.database = database
-        self.collection = collection
-
-    def get(self, id: str) -> T:
-        if not ObjectId.is_valid(id):
-            return None
-        id_obj = ObjectId(id)
-        document = self.database[self.collection].find_one({"_id": id_obj})
-        if document is None:
-            return None
-
-        return from_document(document)
-
-    def update(self, id: str, item: T) -> Tuple[bool, Any]:
-        try:
-            result = self.database[self.collection].update_one({"_id": ObjectId(id)}, {"$set": item})
-            if result.modified_count == 1:
-                document = self.database[self.collection].find_one({"_id": ObjectId(id)})
-                return True, from_document(document)
-            else:
-                return False, "Entity not found"
-        except Exception as e:
-            return False, str(e)
-
-    def add(self, item: T) -> Tuple[bool, Any]:
-        try:
-            result = self.database[self.collection].insert_one(item)
-            id = result.inserted_id
-            document = self.database[self.collection].find_one({"_id": id})
-            return True, from_document(document)
-        except Exception as e:
-            return False, str(e)
-
-    def delete(self, id: str) -> bool:
-        result = self.database[self.collection].delete_one({"_id": ObjectId(id)})
-        return result.deleted_count == 1
-
-    def list(
-        self,
-        limit: int,
-        skip: int,
-        sort_key: str,
-        sort_order=pymongo.DESCENDING,
-        **kwargs,
-    ) -> Dict[int, List[T]]:
-        """List entities
         param limit: The maximum number of entities to return
             type: int
         param skip: The number of entities to skip
@@ -122,26 +75,168 @@ class MongoDBStore(Store[T], Generic[T]):
         param kwargs: Additional query parameters
             type: dict
             example: {"key": "models"}
-        return: A dictionary with the count and the result
+        return: A list of entities
         """
-        cursor = self.database[self.collection].find(kwargs).sort(sort_key, sort_order).skip(skip or 0).limit(limit or 0)
+        pass
 
-        count = self.database[self.collection].count_documents(kwargs)
-
-        result = [from_document(document) for document in cursor]
-
-        return {"count": count, "result": result}
-
+    @abstractmethod
     def count(self, **kwargs) -> int:
-        """Count entities
+        """Count entities.
+
         param kwargs: Additional query parameters
             type: dict
             example: {"key": "models"}
-        return: The count (int)
+        return: The count of entities
         """
+        pass
+
+
+class MongoDBStore(Store):
+    """Base MongoDB store implementation."""
+
+    def __init__(self, database: Database, collection: str, primary_key: str) -> None:
+        """Initialize MongoDBStore."""
+        self.database = database
+        self.collection = collection
+        self.primary_key = primary_key
+        self.database[self.collection].create_index([(self.primary_key, pymongo.DESCENDING)])
+
+    def mongo_get(self, id: str) -> Dict:
+        document = self.database[self.collection].find_one({self.primary_key: id})
+        if document is None:
+            return None
+        return document
+
+    def mongo_add(self, item: Dict) -> Tuple[bool, Any]:
+        try:
+            if self.primary_key not in item or not item[self.primary_key]:
+                item[self.primary_key] = str(uuid.uuid4())
+
+            item["committed_at"] = datetime.now()
+
+            self.database[self.collection].insert_one(item)
+            document = self.database[self.collection].find_one({self.primary_key: item[self.primary_key]})
+            return True, document
+        except Exception as e:
+            return False, str(e)
+
+    def mongo_update(self, item: Dict) -> Tuple[bool, Any]:
+        try:
+            id = item[self.primary_key]
+            result = self.database[self.collection].update_one({self.primary_key: id}, {"$set": item})
+            if result.modified_count == 1:
+                document = self.database[self.collection].find_one({self.primary_key: id})
+                return True, document
+            return False, "Entity not found"
+        except Exception as e:
+            return False, str(e)
+
+    def mongo_delete(self, id: str) -> bool:
+        result = self.database[self.collection].delete_one({self.primary_key: id})
+        return result.deleted_count == 1
+
+    def mongo_select(
+        self,
+        limit: int = 0,
+        skip: int = 0,
+        sort_key: str = None,
+        sort_order=pymongo.DESCENDING,
+        **kwargs,
+    ) -> List[Dict]:
+        _sort_order = sort_order or pymongo.DESCENDING
+        if sort_key and sort_key != self.primary_key:
+            cursor = (
+                self.database[self.collection]
+                .find(kwargs)
+                .sort({sort_key: _sort_order, self.primary_key: pymongo.DESCENDING})
+                .skip(skip or 0)
+                .limit(limit or 0)
+            )
+        else:
+            cursor = self.database[self.collection].find(kwargs).sort(self.primary_key, pymongo.DESCENDING).skip(skip or 0).limit(limit or 0)
+
+        return [document for document in cursor]
+
+    def mongo_count(self, **kwargs) -> int:
         return self.database[self.collection].count_documents(kwargs)
 
 
-class SQLStore(Store[T]):
-    def __init__(self, Session):
+class SQLStore(Generic[T]):
+    """Base SQL store implementation."""
+
+    def __init__(self, Session: Type[SessionClass], SQLModel: Type[T]) -> None:
+        """Initialize SQLStore."""
+        self.SQLModel = SQLModel
         self.Session = Session
+
+    def sql_get(self, session: SessionClass, id: str) -> T:
+        stmt = select(self.SQLModel).where(self.SQLModel.id == id)
+        return session.scalars(stmt).first()
+
+    def sql_add(self, session, entity: T) -> Tuple[bool, Any]:
+        if not entity.id:
+            entity.id = str(uuid.uuid4())
+        session.add(entity)
+        session.commit()
+
+        return True, entity
+
+    def sql_update(self, session: SessionClass, item: Dict) -> Tuple[bool, Any]:
+        stmt = select(self.SQLModel).where(self.SQLModel.id == item["id"])
+        existing_item = session.scalars(stmt).first()
+
+        if existing_item is None:
+            return False, "Item not found"
+        for key, value in item.items():
+            setattr(existing_item, key, value)
+
+        session.commit()
+
+        return True, existing_item
+
+    def sql_delete(self, id: str) -> bool:
+        with self.Session() as session:
+            stmt = select(self.SQLModel).where(self.SQLModel.id == id)
+            item = session.scalars(stmt).first()
+
+            if item is None:
+                return False
+
+            session.delete(item)
+            session.commit()
+
+            return True
+
+    def sql_select(self, session: SessionClass, limit: int = 0, skip: int = 0, sort_key: str = None, sort_order=pymongo.DESCENDING, **kwargs) -> List[T]:
+        stmt = select(self.SQLModel)
+
+        for key, value in kwargs.items():
+            stmt = stmt.where(getattr(self.SQLModel, key) == value)
+
+        _sort_order = sort_order or pymongo.DESCENDING
+
+        secondary_sort_obj = self.SQLModel.__table__.columns.get("id").desc()
+        if sort_key and sort_key in self.SQLModel.__table__.columns:
+            sort_obj = self.SQLModel.__table__.columns.get(sort_key)
+            if _sort_order == pymongo.DESCENDING:
+                sort_obj = sort_obj.desc()
+
+            stmt = stmt.order_by(sort_obj, secondary_sort_obj)
+        else:
+            stmt = stmt.order_by(secondary_sort_obj)
+
+        if limit:
+            stmt = stmt.offset(skip or 0).limit(limit)
+        elif skip:
+            stmt = stmt.offset(skip)
+
+        return session.scalars(stmt).all()
+
+    def sql_count(self, **kwargs) -> int:
+        with self.Session() as session:
+            stmt = select(func.count()).select_from(self.SQLModel)
+
+            for key, value in kwargs.items():
+                stmt = stmt.where(getattr(self.SQLModel, key) == value)
+
+            return session.scalar(stmt)
