+++ fedn/tests/stores/test_combiner_store.py
@@ -115,10 +115,10 @@ class TestCombinerStore:
         for (name1, db_1), (name2, db_2) in zip(db_connections_with_data[1:], db_connections_with_data[:-1]):
             print("Running tests between databases {} and {}".format(name1, name2))
             for *opt,kwargs in options:
-                gathered_combiners = db_1.combiner_store.select(*opt, **kwargs)
+                gathered_combiners = db_1.combiner_store.list(*opt, **kwargs)
                 count =  db_1.combiner_store.count(**kwargs)
                 
-                gathered_combiners2 = db_2.combiner_store.select(*opt, **kwargs)
+                gathered_combiners2 = db_2.combiner_store.list(*opt, **kwargs)
                 count2 = db_2.combiner_store.count(**kwargs)
 
                 assert(count == count2)
