+++ fedn/tests/stores/test_package_store.py
@@ -90,10 +90,10 @@ class TestPackageStore:
         for (name1, db_1), (name2, db_2) in zip(db_connections_with_data[1:], db_connections_with_data[:-1]):
             print("Running tests between databases {} and {}".format(name1, name2))
             for *opt,kwargs in options:
-                gathered_packages = db_1.package_store.select(*opt, **kwargs)
+                gathered_packages = db_1.package_store.list(*opt, **kwargs)
                 count = db_1.package_store.count(**kwargs)
                 
-                gathered_packages2 = db_2.package_store.select(*opt, **kwargs)
+                gathered_packages2 = db_2.package_store.list(*opt, **kwargs)
                 count2 = db_2.package_store.count(**kwargs)
                 
                 assert count == count2
