+++ fedn/tests/stores/test_prediction_store.py
@@ -108,10 +108,10 @@ class TestPredictionStore:
         for (name1, db_1), (name2, db_2) in zip(db_connections_with_data[1:], db_connections_with_data[:-1]):
             print("Running tests between databases {} and {}".format(name1, name2))
             for *opt,kwargs in options:
-                gathered_models = db_1.prediction_store.select(*opt, **kwargs)
+                gathered_models = db_1.prediction_store.list(*opt, **kwargs)
                 count = db_1.prediction_store.count(**kwargs)
 
-                gathered_models2 = db_2.prediction_store.select(*opt, **kwargs)
+                gathered_models2 = db_2.prediction_store.list(*opt, **kwargs)
                 count2 = db_2.prediction_store.count(**kwargs)
 
                 assert count == count2
