+++ fedn/tests/stores/test_round_store.py
@@ -161,10 +161,10 @@ class TestRoundStore:
         for (name1, db_1), (name2, db_2) in zip(db_connections_with_data[1:], db_connections_with_data[:-1]):
             print("Running tests between databases {} and {}".format(name1, name2))
             for *opt,kwargs in options:
-                gathered_rounds = db_1.round_store.select(*opt, **kwargs)
+                gathered_rounds = db_1.round_store.list(*opt, **kwargs)
                 count = db_1.round_store.count(**kwargs)
 
-                gathered_rounds2 = db_2.round_store.select(*opt, **kwargs)
+                gathered_rounds2 = db_2.round_store.list(*opt, **kwargs)
                 count2 = db_2.round_store.count(**kwargs)
                 
                 assert count == count2
