+++ fedn/tests/stores/test_status_store.py
@@ -92,10 +92,10 @@ class TestStatusStore:
         for (name1, db_1), (name2, db_2) in zip(db_connections_with_data[1:], db_connections_with_data[:-1]):
             print("Running tests between databases {} and {}".format(name1, name2))
             for *opt,kwargs in options:
-                gathered_statuses = db_1.status_store.select(*opt, **kwargs)
+                gathered_statuses = db_1.status_store.list(*opt, **kwargs)
                 count =  db_1.status_store.count(**kwargs)
 
-                gathered_statuses2 = db_2.status_store.select(*opt, **kwargs)
+                gathered_statuses2 = db_2.status_store.list(*opt, **kwargs)
                 count2 =  db_2.status_store.count(**kwargs)
 
                 assert count == count2
